+++ src/libasr/string_utils.cpp
@@ -127,21 +127,36 @@ std::string replace(const std::string &s,
     return std::regex_replace(s, std::regex(regex), replace);
 }
 
-std::string read_file(const std::string &filename)
+bool read_file(const std::string &filename, std::string &text)
 {
+    if (filename.empty()) return false;
     std::ifstream ifs(filename.c_str(), std::ios::in | std::ios::binary
             | std::ios::ate);
+    if (!ifs.is_open()) return false;
 
     std::ifstream::pos_type filesize = ifs.tellg();
-    if (filesize < 0) return std::string();
+    if (filesize < 0) return false;
 
     ifs.seekg(0, std::ios::beg);
 
     std::vector<char> bytes(filesize);
+    if (filesize == 0) bytes.reserve(1);
     ifs.read(&bytes[0], filesize);
 
-    return std::string(&bytes[0], filesize);
+    text = std::string(&bytes[0], filesize);
+    return true;
 }
+
+std::string read_file_ok(const std::string &filename) {
+    std::string text;
+    if (read_file(filename, text)) {
+        return text;
+    } else {
+        std::cerr << "File '" + filename + "' cannot be opened." << std::endl;
+        abort();
+    }
+}
+
 
 std::string parent_path(const std::string &path) {
     int pos = path.size()-1;
