+++ src/libasr/string_utils.h
@@ -29,7 +29,10 @@ char *s2c(Allocator &al, const std::string &s);
 std::string replace(const std::string &s,
     const std::string &regex, const std::string &replace);
 
-std::string read_file(const std::string &filename);
+// Reads a file, returns success/fail as a result
+bool read_file(const std::string &filename, std::string &text);
+// Reads a file, aborts on failure
+std::string read_file_ok(const std::string &filename);
 
 // Returns the parent path to the given path
 std::string parent_path(const std::string &path);
