+++ src/libasr/utils.h
@@ -141,7 +141,6 @@ struct CompilerOptions {
     CompilerOptions () : platform{get_platform()} {};
 };
 
-bool read_file(const std::string &filename, std::string &text);
 bool present(Vec<char*> &v, const char* name);
 bool present(char** const v, size_t n, const std::string name);
 int initialize();
